Add an example for EventReg
