+++ src/Event/Event.h
@@ -1,0 +1,3 @@
+#pragma once
+
+namespace ll::event::inline TSETEvent {} // namespace ll::event::inline TSETEvent
