+++ src/Event/TPAEvent/TPARequestEvent.cpp
@@ -1,0 +1,11 @@
+#include "Event/TPAEvent/TPARequestEvent.h"
+#include "mc/world/actor/player/Player.h"
+
+namespace TSETEvent::inline TPAEvent {
+
+Player& TPARequestEvent::getSender() { return *mSender; };
+
+Player& TPARequestEvent::getTarget() { return *mTarget; };
+
+
+}
