--- src/Events/Events.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "ll/api/event/Event.h"
-
-namespace TSETEvent {
-} // namespace TSETEvent
