--- src/Events/TPAEvents/TPARequestEvent.h
@@ -1,13 +1,0 @@
-#include "ll/api/event/Event.h"
-
-// 写你妈
-
-namespace TSETEvent {
-class TPARequestEvent final : public ll::event::Cancellable<ll::event::Event> {
-private:
-    Player* player;
-    Player* taget;
-public:
-    // 算了，你自己写吧
-};
-} // namespace TSETEvent
